+++ Avalonia.Lottie/Animation/Keyframe/ValueCallbackKeyframeAnimation.cs
@@ -16,7 +16,7 @@ namespace Avalonia.Lottie.Animation.Keyframe
         ///     If this doesn't return 1, then <see cref="set_Progress" /> will always clamp the progress
         ///     to 0.
         /// </summary>
-        protected override float EndProgress => 1f;
+        protected override double  EndProgress => 1f;
 
         public override TA Value =>
             ValueCallback.GetValueInternal(0f, 0f, default, default, Progress, Progress, Progress);
@@ -26,7 +26,7 @@ namespace Avalonia.Lottie.Animation.Keyframe
             if (ValueCallback != null) base.OnValueChanged();
         }
 
-        public override TA GetValue(Keyframe<TK> keyframe, float keyframeProgress)
+        public override TA GetValue(Keyframe<TK> keyframe, double  keyframeProgress)
         {
             return Value;
         }
