+++ Avalonia.Lottie/MathExt.cs
@@ -9,7 +9,7 @@ namespace Avalonia.Lottie
             return Math.Sqrt(x * x + y * y);
         }
 
-        public static float Lerp(float value1, float value2, float amount)
+        public static double  Lerp(double value1, double  value2, double  amount)
         {
             return value1 + (value2 - value1) * amount;
         }
