+++ Avalonia.Lottie/Value/LottieValueCallback.cs
@@ -38,13 +38,13 @@ namespace Avalonia.Lottie.Value
         }
 
         public T GetValueInternal(
-            float startFrame,
-            float endFrame,
+            double  startFrame,
+            double  endFrame,
             T startValue,
             T endValue,
-            float linearKeyframeProgress,
-            float interpolatedKeyframeProgress,
-            float overallProgress
+            double  linearKeyframeProgress,
+            double  interpolatedKeyframeProgress,
+            double  overallProgress
         )
         {
             return GetValue(
